+++ gui/components/__init__.py
@@ -1,5 +1,9 @@
 from .analysis import AnalysisParamsCard
+from .analysis_utils import analysis_label, present_timestamp
+from .choice_fork import two_button_choice_fork_content
+from .import_options import ImportOptionsDialog
 from .manage_analyses import ManageAnalysisDialog
+from .manage_projects import ManageProjectsDialog
 from .stepper_steps import (
     AnalyzerSelectionStep,
     ColumnMappingStep,
@@ -15,8 +19,13 @@ __all__ = [
     "AnalysisParamsCard",
     "UploadButton",
     "ManageAnalysisDialog",
+    "ManageProjectsDialog",
+    "ImportOptionsDialog",
     "AnalyzerSelectionStep",
     "ColumnMappingStep",
     "ParamsConfigStep",
     "RunAnalysisStep",
+    "analysis_label",
+    "present_timestamp",
+    "two_button_choice_fork_content",
 ]
