+++ gui/pages/analyzer_select.py
@@ -1,5 +1,5 @@
 from gui.base import GuiPage
-from gui.choice_fork import two_button_choice_fork_content
+from gui.components.choice_fork import two_button_choice_fork_content
 from gui.routes import gui_routes
 from gui.session import GuiSession
